hopefully fixed rust code

+++ vox_core/src/gdres.rs
@@ -6,17 +6,17 @@ use godot::classes::Resource;
 #[derive(GodotClass, Debug, Default)]
 #[class(base=Resource, init)]
 pub struct GdRes {
-    hash: u64,
+    hash: u32,
     resource: Gd<Resource>
 }
 
 impl GdRes {
 
-    pub fn new(hash: u64, resource: Gd<Resource>) -> GdRes {
+    pub fn new(hash: u32, resource: Gd<Resource>) -> GdRes {
        GdRes { hash, resource }
     }
 
-    pub fn get_hash(&self) -> u64 {
+    pub fn get_hash(&self) -> u32 {
         self.hash
     }
